Replace dependency `elixir-uuid` with `uniq`
This [thread](https://github.com/zyro/elixir-uuid/issues/50) on the
`elixir-uuid` indicates that the project is no longer maintained. The
successor project is the [uniq](https://hex.pm/packages/uniq).

This
[documentation](https://hexdocs.pm/uniq/migration_from_elixir_uuid.html)
describes how to migrate from `elixir-uuid` to `uniq`, which its very
simple to do.

The `elixir-uuid` dependency also has a compilation warning with elixir
1.14 of `warning: use Bitwise is deprecated. import Bitwise instead`.